% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TcGSA-internal.R
\name{TcGSA-internal}
\alias{MYheatmap.2}
\alias{STSdist}
\alias{TcGSA-internal}
\alias{TcGSA.dataLME}
\alias{TcGSA.formula}
\title{Internal TcGSA Functions}
\description{
Internal TcGSA functions
}
\details{
These are not to be called by the user.
}
\keyword{internal}

