% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TcGSA-internal.R
\name{TcGSA-internal}
\alias{TcGSA-internal}
\alias{MYheatmap.2}
\alias{TcGSA.formula}
\alias{TcGSA.dataLME}
\alias{STSdist}
\title{Internal TcGSA Functions}
\description{
Internal TcGSA functions
}
\details{
These are not to be called by the user.

\code{MYheatmap.2} was adapted from the function \code{\link{heatmap.2}} 
included in the package \code{gplots}, authored by: 
Andy Liaw, original; 
R. Gentleman, M. Maechler, W. Huber, G. Warnes, revisions;
under the GPL-2 license.
}
\keyword{internal}
