% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TcGSA-internal.R
\name{TcGSA-internal}
\alias{MYheatmap.2}
\alias{STSdist}
\alias{TcGSA-internal}
\alias{TcGSA.dataLME}
\alias{TcGSA.formula}
\title{Internal TcGSA Functions}
\description{
Internal TcGSA functions
}
\details{
These are not to be called by the user.

MYheatmap.2 was adapted from the function 'heatmap.2' included in the package 'gplots'
(see \code{\link[gplots:heatmap.2]{heatmap.2}}), authored by:
Andy Liaw, original;
R. Gentleman, M. Maechler, W. Huber, G. Warnes, revisions;
under the GPL-2 license.
}
\keyword{internal}

