% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TcGSA-package.R
\docType{package}
\name{TcGSA-package}
\alias{TcGSA}
\alias{TcGSA-package}
\title{Time-course Gene Set Analysis}
\description{
This package implements TcGSA, an algorithm to analyze longitudinal gene-expression data at the gene set level.
}
\details{
\tabular{ll}{
Package: \tab TcGSA\cr
Type: \tab Package\cr
Version: \tab 0.9.91\cr
Date: \tab 2015-03-11\cr
License:\tab \href{http://www.gnu.org/licenses/lgpl.txt}{LGPL-3}\cr
}
The main function in this package is \code{\link{TcGSA.LR}} which performs Time-course Gene Set Analysis, and provide nice representations of its results (see \code{\link{plot.TcGSA}} and \code{\link{plot1GS}}).
}
\author{
Boris P. Hejblum

Damien Chimits
--- Maintainer: Boris P. Hejblum
}
\references{
Hejblum, B.P., Skinner, J., Thiebaut, R., 2014, TcGSA: a gene set approach for longitudinal gene expression data analysis, \bold{submitted}.
}
\seealso{
\code{\link[GSA:GSA]{GSA}}
}

