\name{TcGSA-package}
\alias{TcGSA-package}
\alias{TcGSA}

\docType{package}

\title{
Time-course Gene Set Analysis
}

\description{
This package implements TcGSA, an algorithm to analyze longitudinal gene-expression data at the gene set level.
}

\details{
\tabular{ll}{
Package: \tab TcGSA\cr
Type: \tab Package\cr
Version: \tab 0.9.14\cr
Date: \tab 2013-01-14\cr
License: \tab \href{http://www.gnu.org/licenses/lgpl.txt}{LGPL-3}\cr
}

The main function in this package is \code{\link{TcGSA.LR}} which performs Time-course Gene Set Analysis, and has a nice representation of its results (see \code{\link{plot.TcGSA}}).

}

\author{
Boris P. Hejblum

Maintainer: Boris P. Hejblum <boris.hejblum@isped.u-bordeaux2.fr>
}

\references{
Hejblum, B.P., Skinner, J., Thiebaut, R., 2012, TcGSA: Time-course gene set analysis of longitudinal gene expression, with an application to a therapeutic HIV vaccine trial, \bold{submitted}.
}
\keyword{package}

\seealso{
\code{\link[GSA:GSA]{GSA}}
}

