% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TcGSA-package.R
\docType{package}
\name{TcGSA-package}
\alias{TcGSA-package}
\alias{TcGSA}
\title{Time-course Gene Set Analysis}
\description{
This package implements TcGSA, an algorithm to analyze longitudinal gene-expression data at the gene set level.
}
\details{
\tabular{ll}{
Package: \tab TcGSA\cr
Type: \tab Package\cr
Version: \tab 0.12.9\cr
Date: \tab 2022-02-27\cr
License:\tab \href{http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt}{GPL-2}\cr
}
The main function in this package is \code{\link{TcGSA.LR}} which performs Time-course Gene Set Analysis, and provide nice representations of its results (see \code{\link{plot.TcGSA}} and \code{\link{plot1GS}}).
}
\references{
Hejblum BP, Skinner J, Thiebaut R, (2015)
Time-Course Gene Set Analysis for Longitudinal Gene Expression Data.
\emph{PLOS Comput. Biol.} 11(6):e1004310.
doi: 10.1371/journal.pcbi.1004310
}
\seealso{
\code{\link[GSA:GSA]{GSA}} \code{\link[dearseq:dgsa_seq]{dgsa_seq}}
}
\author{
Boris P. Hejblum, Damien Chimits
--- Maintainer: Boris P. Hejblum
}
